pyproject.toml
[project]
name = "dbt-dagsterizer"
version = "0.1.11"
description = "Dagster automation driven by dbt metadata (Luban CI)"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "click>=8.3.1",
    "cookiecutter>=2.6.0",
    "dagster>=1.12,<2",
    "dagster-dbt>=0.26,<1",
    "dbt-core>=1.9,<2",
    "dbt-starrocks>=1.9,<2",
    "pymysql>=1.1.0",
    "ruamel-yaml>=0.19.1",
]

[tool.hatch.build.targets.wheel]
packages = ["src/dbt_dagsterizer"]

[tool.hatch.build.targets.wheel.shared-data]
"src/dbt_dagsterizer/project_templates" = "dbt_dagsterizer/project_templates"

[dependency-groups]
dev = [
    "pytest>=8.0.0",
    "ruff>=0.12.0",
]

[project.scripts]
dbt-dagsterizer = "dbt_dagsterizer.cli:cli"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
target-version = "py312"
line-length = 100
extend-exclude = [
    "src/dbt_dagsterizer/project_templates",
]

[tool.ruff.lint]
select = ["E", "F", "I"]
ignore = ["E501"]

[tool.pytest.ini_options]
testpaths = ["tests"]
